+++ OpenSim/Actuators/Test/testDeGrooteFregly2016Muscle.cpp
@@ -142,9 +142,8 @@ TEST_CASE("DeGrooteFregly2016Muscle basics") {
     musclePtr->set_ignore_tendon_compliance(true);
     musclePtr->set_fiber_damping(0);
     musclePtr->setName("muscle");
-    auto& path = musclePtr->updPath<GeometryPath>();
-    path.appendNewPathPoint("origin", model.updGround(), SimTK::Vec3(0));
-    path.appendNewPathPoint("insertion", *body, SimTK::Vec3(0));
+    musclePtr->addNewPathPoint("origin", model.updGround(), SimTK::Vec3(0));
+    musclePtr->addNewPathPoint("insertion", *body, SimTK::Vec3(0));
     model.addComponent(musclePtr);
     auto& muscle = model.updComponent<DeGrooteFregly2016Muscle>("muscle");